This project Done

+++ Serverside/Src/Controllers/notes.ts
@@ -2,10 +2,12 @@ import { RequestHandler } from "express";
 import NoteModel from "../Models/note";
 import mongoose from "mongoose";
 import createHttpError from "http-errors";
+import { assertlsDefined } from "../util/assertlsDefined";
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authenticatedUserId = req.session.userId;
   try {
-    // throw createHttpError(401);
-    const notes = await NoteModel.find().exec();
+    assertlsDefined(authenticatedUserId);
+    const notes = await NoteModel.find({ userId: authenticatedUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -13,10 +15,10 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 };
 
 export const getNote: RequestHandler = async (req, res, next) => {
-  // let id = mongoose.Types.ObjectId(id);
-  //console.log(noteId);
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertlsDefined(authenticatedUserId);
     const objectId = new mongoose.Types.ObjectId(noteId);
     if (!mongoose.isValidObjectId(objectId)) {
       throw createHttpError(400, "Invalid note id");
@@ -25,8 +27,11 @@ export const getNote: RequestHandler = async (req, res, next) => {
     if (!note) {
       throw createHttpError(404, "Note not found");
     }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
+    }
+
     res.status(200).json(note);
-    //console.log(note);
   } catch (error) {
     next(error);
   }
@@ -45,11 +50,14 @@ export const createNotes: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertlsDefined(authenticatedUserId);
     if (!title) {
       throw createHttpError(400, "Note Must have a Title!..");
     }
     const newNote = await NoteModel.create({
+      userId: authenticatedUserId,
       title: title,
       text: text,
     });
@@ -77,8 +85,9 @@ export const updateNote: RequestHandler<
   const noteId = req.params.noteId;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authenticatedUserId = req.session.userId;
   try {
-    // const objectId = new mongoose.Types.ObjectId(noteId);
+    assertlsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
@@ -88,6 +97,9 @@ export const updateNote: RequestHandler<
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not Found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
     note.title = newTitle;
     note.text = newText;
@@ -100,13 +112,18 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const noteId = req.params.noteId;
+  const authenticatedUserId = req.session.userId;
   try {
+    assertlsDefined(authenticatedUserId);
     if (!mongoose.isValidObjectId(noteId)) {
       throw createHttpError(400, "Invalid note id");
     }
     const note = await NoteModel.findById(noteId).exec();
     if (!note) {
       throw createHttpError(404, "Note not found");
+    }
+    if (!note.userId.equals(authenticatedUserId)) {
+      throw createHttpError(401, "You cannot access this note");
     }
     await note.deleteOne();
     res.sendStatus(204);
